renaming exports and importing Play and Garage from index.js in containers directory
