+++ client/src/Root.js
@@ -3,9 +3,11 @@ import './Root.scss';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
 import Main from './containers/Main';
-import Garage from './containers/Garage';
-import Game from './containers/Play/Play';
+// import Garage from './containers/Garage/Garage';
+// import Game from './containers/Play/Play';
 import Shop from './containers/Shop';
+
+import { Garage, Game } from './containers';
 
 function Root() {
   return (
