+++ client/src/containers/index.js
@@ -1,1 +1,2 @@
-export { default as Play } from './Play/Play';
+export { default as Game } from './Play/Play';
+export { default as Garage } from './Garage/Garage';
